Refactored mixin out of PostListView stuff, better default order
For #19
